1.解决分组的bug 2.groupSize重新设置为uint16_t

+++ src/primitives/block.h
@@ -92,7 +92,7 @@ class MCBlock : public MCBlockHeader
 public:
     // network and disk
     std::vector<MCTransactionRef> vtx;
-    std::vector<uint8_t> groupSize;
+    std::vector<uint16_t> groupSize;
     std::vector<ContractPrevData> prevContractData;
 
     // memory only
